addition to 'parent was fixed for johnThread', it did not work
Addition to 8f707af0a02a76b751efe467f2c1c6d45ef3847c

Parent was used in a different method. So in that method parent is a
method inherited from QObject. It is needed to call it to get a
pointer to the parent. Again the problem caused by commit without any
testing.